refactor: separate contexts
