+++ client/src/contexts/index.ts
@@ -1,0 +1,6 @@
+import UserInfoProvider, { UserContext } from './userInfo';
+import SnackBarProvider, { SnackBarContext } from './snackBar';
+
+export {
+  UserInfoProvider, UserContext, SnackBarProvider, SnackBarContext,
+};
